Modernize configure.ac with autoupdate
Take autoupdate's advice to remove obsolete tests.

It recommends depending on autoconf 2.71 but this is wound back to 2.69
as that's the latest Debian stable release. It can probably be rewound
further.

Also, tests we don't use are removed. There's no need to have them if
we won't use them, as we can't rely on them in code unless CMake also
sets the same tests in the same circumstances.

Finally, add AM_SILENT_RULES to make it easier to spot compilation
warnings.