Use conditional build objects rather than subdirs
Structing the build of the the native device driver as selecting a subdir
based on a conditional variable does not let automake recognise that all
three subdirs need to be included in the distribution. It doesn't even
know what those subdirs are, they're only defined in configure.ac which
it can't read. What works better is to define three conditionals that
each can add one native driver as a conditional library. This also
removes the need to link multiple libraries in all places, so we don't
need the ADFLIB_OS conditional.