Fix scripts so they run during 'make distcheck'
Scripts need to run from the build directory, which contains no source
files. They need to look to $srcdir for their source files.

A special case was needed for floppy.sh (mirrored in bigdev.sh) because
it wants to have a list of bare filenames (some scripts, some programs)
and then just execute each one in the list as if they're all in the
same directory (they're not). So instead we need to test on each
iteration of the loop which directory they're in.

Also, just to check that srcdir can be on read-only media, distcheck
marks all srcdir files as read only. These scripts copy a read-only ADF
dump to the build directory to modify it, using the "cp" command which
preserves the file permissions, so then all the tests fail on this
unexpectedly read-only file. Use "cat src >dest"