Make Win32 filename mangling optional with unadf -w
Add a new command-line flag, -w, for unadf. If provided, it will turn on
filename mangling. If not, filenames will be extracted as-is, which is
not a problem for UNIX filesystems, they can accomodate all Amiga
filenames.

Extend the mangling so it supports fixing all forbidden filenames
in Windows environments.