Only have a single adf_nativ.h
Only have a single adf_nativ.h which defines struct AdfNativeFunctions
and declares adfInitNativeFct()

Each implementation of adf_nativ.c is then self-contained, and has no
headers that affect the rest of the library

To be sure, make all the functions in the adf_nativ.c files static,
except for adfInitNativeFct(), which now becomes the last function
in the file to remove the need for prototypes.

As src/generic/adf_nativ.c does nothing, and lies with RC_OK
when it didn't actually read/write anything, make its adfIsDevNative()
always return FALSE, so it never get invoked. Previously it claimed to
handle any file under /dev/, which it didn't.