Fixed potential access to uninitialized values in sectList
I think this is a bit overzealous from the Xcode analyzer but unless there is a reason to call adfReadRootBlock() where it’s currently called, it does receive a non-const pointer to vol and the code releis on vol->bitmap.size for a lot of its acesses to sectList. There is here nothing guarding against this value changing during the call to adfReadRootBlock().

This guarantees that the value used is constant throughout.