examples: reorganize tests.
Goals:
1. Separate expected output from scripts.
2. Make the tests work with MSYS2.

1. is required by 2. since under MSYS2 the output differs
largely (EOL, paths). Mangling these in a script is maybe possible,
but inherently problematic. The biggest problem comes from
DOS/Windows-style paths using '\' as separator, which is also
the escape character in shell scripts...
Consequence: the escape characters are working somewhat differently
on MSYS2 and on Linux and creating a common script is absurdly hard.
Separating test outputs allows to use different files with expected
contents depending on the platform.

Note on unadf tests:
file permission checks are not working on MSYS2 - either the way
of checking or permissions are invalid.
(File permission test is disabled in the uadf test script).