adfMountFlop: bugfix (incorrectly set vol->mounted).
vol->mounted remained set in case of error while reading
root block - what lead to an invalid state of vol.
Changed to unsetting it immediately after trying to read
the root block.
Also - setting it false by default (it was not initialized
properly in case the filesystem was not present or was
not a DOS filesystem (OFS/FFS).