Refactoring: simplify and standardize safe duplicating strings from raw block data.
Note that 'safe' concerns the size, not contents. The string can still
contain invalid / binary data, which, for instance, if formatted in a way
to perform an attack, can cause harm if outputted to a terminal.

This has to be taken care in printing functions, as internally consistency
with raw data from blocks must be kept.