adfDevMount: improve checking and switching the hard disk/file dev. type.
The device types HARDFILE and HARDDISK are used interchangably,
depending if the device is partitioned (contains an RDSK block etc.)
or not. Calling a create device function (which sets up the volume /
partition structure), the type can be changed, so the mount function
should be able to detect properly the type, and switch it both ways.

The priority is changed to start from detecting the presence of
an RDSK block, and, if it is not found, treat the device as
a HARDFILE, an unpartitioned device, with a single volume using
the whole device (like a floppy disk).