adfVolMount: remove setting fstype.
Volume's fstype must be already set by adfDevMount or adfDevCreate.
No need to read and parse the bootblock - besides filesystem
signature, it is not needed for accessing filesystem data.

Note that the requirement to load a valid bootblock is a problem
in case of many non-bootable disks (even Workbench Extras),
which do not have a valid bootblock, with correct checksums etc.,
but only the 'DOSn' signature at the beginning. Mounting such
volumes was failing with a checksum error on reading bootblock,
if 'ignoreChecksumErrors' was not set.

This change alleviates the issue.