adf_salv / adfUndelFile: bugfix (mark also the file header block as used).
While files where undeleted successfully (visible, readable),
there was an inconsistency in the volume's block allocation
bitmap: the file header block of the recovered file was still
marked as not used. The result was that the adf_salvage utility
was still showing the file as 'deleted'.

This would also result in inevitable corruption / data loss
in case of allocating new blocks for writing - the file header
block could be allocated (as a 'free' one) and overwritten with
a block of another entry.