src / autotools: make the private headers private.
Library headers listed as _HEADERS are being installed along
with the library, while the private sources should not
(they are needed only for building the library itself).
Such private headers can be specified as noist_HEADERS, or,
what is recommender by Automake manual, just as _SOURCES.

Note, that if the private headers are completely omitted,
make distcheck fails.